requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
requests
